requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
flask>=3.0.0
gunicorn>=21.2.0

# ── Google Cloud / AI ───────────────────────────────────────────────────────��─
google-cloud-aiplatform>=1.60.0         # Vertex AI SDK (includes RAG Engine)
google-cloud-bigquery>=3.25.0           # BigQuery client
google-cloud-bigquery-storage>=2.25.0   # BigQuery Storage API (fast reads)
google-cloud-storage>=2.18.0            # GCS client
google-cloud-pubsub>=2.22.0             # Pub/Sub
google-generativeai>=0.8.0              # Gemini API
google-genai>=1.5.0                     # google.genai client
google-adk>=0.5.0                       # Agent Development Kit

# ── Data / ML ─────────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0

# ── HTTP / Networking ─────────────────────────────────────────────────────────
requests>=2.31.0

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.0

# ── CLI / Display ─────────────────────────────────────────────────────────────
rich>=13.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
faker>=24.0.0
